// apps/web/src/types/database.ts
export type VehicleCategory =
  | "citadine"
  | "compacte"
  | "berline"
  | "suv"
  | "break"
  | "monospace"
  | "cabriolet"
  | "coupe"
  | "utilitaire"
  | "minibus"
  | "prestige"
  | "sans_permis";
export type TransmissionType = "manuelle" | "automatique";
export type FuelType =
  "essence" | "diesel" | "hybride" | "hybride_rechargeable" | "electrique" | "gpl";